Deleting unecessary files and added implementations and tests to visitor spec and file. Allowing visitor to see their list of checked out books. Tests passing.
