+++ lib/data.yml
@@ -2,8 +2,8 @@
 - :item:
     :title: Alfons och soldatpappan
     :author: Gunilla Bergström
-  :available: true
-  :return_date: 
+  :available: false
+  :return_date: 14/07/20
 - :item:
     :title: Skratta lagom! Sa pappa Åberg
     :author: Gunilla Bergström
